+++ extractor.py
@@ -2,9 +2,6 @@
 
 import pandas as pd
 
-import os
-import glob
-
 
 def import_data(paths, header=None):
     data = [pd.read_csv(file, header=header) for file in paths]
@@ -12,7 +9,7 @@ def import_data(paths, header=None):
     if not sum([len(p.columns) != fields_len for p in data]):
         data = pd.concat(data, ignore_index=True)
         data = data.drop_duplicates(ignore_index=True)
-        if header == None:
+        if header is None:
             if len(data.columns) == 10:  # products
                 data.columns = [
                     "product_id",
@@ -41,9 +38,9 @@ def import_data(paths, header=None):
         raise Exception("different columns length")
 
 
-def extract_dataset(data, name):
+def extract_dataset(data, name: str):
     if name not in ("product", "variant", "price"):
-        return "Error: invalid name"
+        raise ValueError("invalid name")
     if name == "product":
         fields_name = [
             "product_id",
@@ -82,3 +79,22 @@ def extract_dataset(data, name):
         .drop_duplicates(fields_unique)
         .sort_values(fields_sort, ignore_index=True)
     )
+
+def append_dataset(dataset, master: str, dup=False, dup_name=None, fmt="parquet"):
+    if fmt == "parquet":
+        engine = "fastparquet"
+        master_dataset = pd.read_parquet(master, engine=engine)
+        if dataset.columns.to_list() != master_dataset.columns.to_list():
+            raise Exception("column is different")
+        dataset = pd.concat([master_dataset, dataset]).drop_duplicates(
+            ignore_index=True
+        )
+        if not dup:
+            dup_name = master
+        else:
+            if dup_name is None:
+                dup_name = f"{master.replace('.parquet', '')}_copy.parquet"
+        dataset.to_parquet(dup_name, engine=engine)
+        print(f"save data to {dup_name}")
+    else:
+        raise Exception("not parquet?")
